+++ scripts/install/bazarr.sh
@@ -10,15 +10,36 @@
 #   including (via compiler) GPL-licensed code must also be made available
 #   under the GPL along with build & install instructions.
 
+if [[ -f /tmp/.install.lock ]]; then
+  log="/root/logs/install.log"
+else
+  log="/root/logs/swizzin.log"
+fi
+
+codename=$(lsb_release -cs)
+
 user=$(cut -d: -f1 < /root/.master.info )
-apt-get -y -q install python3-pip python3-dev > /dev/null 2>&1
+if [[ $codename =~ ("bionic"|"stretch"|"xenial"|"jessie") ]]; then
+  . /etc/swizzin/sources/functions/pyenv
+  pyenv_install
+  pyenv_install_version 3.7.7
+  pyenv_create_venv 3.7.7 /home/${user}/.venv/bazarr
+  chown -R ${user}: /home/${user}/.venv/bazarr
+else
+  apt-get -y -q install python3-pip python3-dev python3-venv > $log 2>&1
+  mkdir -p /home/${user}/.venv/bazarr
+  python3 -m venv /home/${user}/.venv/bazarr
+  chown -R ${user}: /home/${user}/.venv/bazarr
+fi
+
 cd /home/${user}
-echo "Cloning into '/home/${user}/bazarr'"
-git clone https://github.com/morpheus65535/bazarr.git > /dev/null 2>&1
+
+echo "Clonando a '/home/${user}/bazarr'"
+git clone https://github.com/morpheus65535/bazarr.git > $log 2>&1
 chown -R ${user}: bazarr
 cd bazarr
-echo "Checking python depends"
-sudo -u ${user} bash -c "pip3 install --user -r requirements.txt" > /dev/null 2>&1
+echo "Comprobando dependencias de pythons"
+sudo -u ${user} bash -c "/home/${user}/.venv/bazarr/bin/pip3 install --user -r requirements.txt" > $log 2>&1
 mkdir -p /home/${user}/bazarr/data/config/
 
 if [[ -f /install/.sonarr.lock ]]; then
@@ -59,7 +80,7 @@ if [[ -f /install/.nginx.lock ]]; then
   sleep 10
   bash /usr/local/bin/swizzin/nginx/bazarr.sh
   service nginx reload
-  echo "Please ensure during bazarr wizard that baseurl is set to: /bazarr/"
+  echo "Por favor asegúrate que durante el asistente de bazarr la baseurl está establecida en: /bazarr/"
 else
 
 cat >> /home/${user}/bazarr/data/config/config.ini <<BAZC
@@ -94,7 +115,7 @@ UMask=0002
 Restart=on-failure
 RestartSec=5
 Type=simple
-ExecStart=/usr/bin/python3 /home/${user}/bazarr/bazarr.py
+ExecStart=/home/${user}/.venv/bazarr/bin/python3 /home/${user}/bazarr/bazarr.py
 KillSignal=SIGINT
 TimeoutStopSec=20
 SyslogIdentifier=bazarr.${user}