+++ scripts/install/deluge.sh
@@ -208,7 +208,7 @@ After=network.target
 
 [Service]
 Type=simple
-User=%I
+User=%i
 
 ExecStart=/usr/bin/deluged -d
 ExecStop=/usr/bin/killall -w -s 9 /usr/bin/deluged
@@ -227,7 +227,7 @@ After=network.target
 
 [Service]
 Type=simple
-User=%I
+User=%i
 
 ExecStart=/usr/bin/deluge-web${args}
 ExecStop=/usr/bin/killall -w -s 9 /usr/bin/deluge-web
@@ -256,7 +256,7 @@ fi
 if [[ -f /tmp/.install.lock ]]; then
   export log="/root/logs/install.log"
 else
-  export log="/dev/null"
+  export log="/root/logs/swizzin.log"
 fi
 local_packages=/usr/local/bin/swizzin
 users=($(cut -d: -f1 < /etc/htpasswd))