+++ scripts/install/flood.sh
@@ -12,10 +12,8 @@ fi
 
 if [[ -f /tmp/.install.lock ]]; then
   log="/root/logs/install.log"
-elif [[ -f /install/.panel.lock ]]; then
-  log="/srv/panel/db/output.log"
 else
-  log="/dev/null"
+  log="/root/logs/swizzin.log"
 fi
 
 . /etc/swizzin/sources/functions/npm
@@ -31,10 +29,10 @@ Description=Flood rTorrent Web UI
 After=network.target
 
 [Service]
-User=%I
-Group=%I
-WorkingDirectory=/home/%I/.flood
-ExecStart=/usr/bin/npm start --production /home/%I/.flood
+User=%i
+Group=%i
+WorkingDirectory=/home/%i/.flood
+ExecStart=/usr/bin/npm start --production /home/%i/.flood
 
 
 [Install]