+++ scripts/remove/quota.sh
@@ -2,19 +2,10 @@
 #
 # [Servidor HD :: Remove quotas]
 #
-# Author             :   liara
-#
-# Servidor HD Copyright (C) 2019
-# Licensed under GNU General Public License v3.0 GPL-3 (in short)
-#
-#   You may copy, distribute and modify the software as long as you track
-#   changes/dates in source files. Any modifications to our software
-#   including (via compiler) GPL-licensed code must also be made available
-#   under the GPL along with build & install instructions.
-#
 #################################################################################
 username=$(cut -d: -f1 < /root/.master.info)
 
 sed -i 's/,usrjquota=aquota.user,jqfmt=vfsv1//g' /etc/fstab
 apt-get remove -y -q quota >/dev/null 2>&1
+rm /etc/sudoers.d/quota
 rm /install/.quota.lock