+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+
+## [2.0.0]
+
+### 19 Abril, 2020
+
+### Cambios
+Introducción del panel Servidor HD dashboard basado en Swizzin Dashboard.
+Añado Jellyfin
+Añado LibreSpeed
+Algunos arreglos y actualizaciones a python
+
 ## [1.7.0]
 
 ### 10 Febrero 2020